requirements.txt
# pip install - r requirements.text || pip3 install -r requirements.txt
playwright
python-dotenv